refactored chunkJournal, journalWriter
