+++ go/store/nbs/file_table_persister.go
@@ -241,3 +241,7 @@ func (ftp *fsTablePersister) PruneTableFiles(ctx context.Context, contents manif
 
 	return nil
 }
+
+func (ftp *fsTablePersister) Close() error {
+	return nil
+}
