+++ go/store/nbs/table_persister.go
@@ -27,6 +27,7 @@ import (
 	"crypto/sha512"
 	"encoding/binary"
 	"errors"
+	"io"
 	"sort"
 )
 
@@ -51,6 +52,8 @@ type tablePersister interface {
 
 	// PruneTableFiles deletes old table files that are no longer referenced in the manifest.
 	PruneTableFiles(ctx context.Context, contents manifestContents) error
+
+	io.Closer
 }
 
 type chunkSourcesByAscendingCount struct {
